+++ cli/azd/cmd/auth_login.go
@@ -13,6 +13,7 @@ import (
 	"strconv"
 	"strings"
 
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/policy"
 	"github.com/MakeNowJust/heredoc/v2"
 	"github.com/azure/azure-dev/cli/azd/cmd/actions"
@@ -239,6 +240,8 @@ func newLoginAction(
 	}
 }
 
+const cLoginSuccessMessage = "Logged in to Azure."
+
 func (la *loginAction) Run(ctx context.Context) (*actions.ActionResult, error) {
 	if len(la.flags.scopes) == 0 {
 		la.flags.scopes = auth.LoginScopes
@@ -254,81 +257,109 @@ func (la *loginAction) Run(ctx context.Context) (*actions.ActionResult, error) {
 			"Next time use `azd auth login`.")
 	}
 
-	if !la.flags.onlyCheckStatus {
-		if err := la.accountSubManager.ClearSubscriptions(ctx); err != nil {
-			log.Printf("failed clearing subscriptions: %v", err)
+	if la.flags.onlyCheckStatus {
+		// In check status mode, we always print the final status to stdout.
+		// We print any non-setup related errors to stderr.
+		// We always return a zero exit code.
+		token, err := la.verifyLoggedIn(ctx)
+		var loginExpiryError *auth.ReLoginRequiredError
+		if err != nil &&
+			!errors.Is(err, auth.ErrNoCurrentUser) &&
+			!errors.As(err, &loginExpiryError) {
+			fmt.Fprintln(la.console.Handles().Stderr, err.Error())
 		}
 
-		if err := la.login(ctx); err != nil {
-			return nil, err
-		}
-	}
-
-	res := contracts.LoginResult{}
-
-	credOptions := auth.CredentialForCurrentUserOptions{
-		TenantID: la.flags.tenantID,
-	}
-
-	if cred, err := la.authManager.CredentialForCurrentUser(ctx, &credOptions); errors.Is(err, auth.ErrNoCurrentUser) {
-		res.Status = contracts.LoginStatusUnauthenticated
-	} else if err != nil {
-		return nil, fmt.Errorf("checking auth status: %w", err)
-	} else {
-		// Ensure ID token is valid, and can be exchanged for an access token
-		token, err := cred.GetToken(ctx, policy.TokenRequestOptions{
-			Scopes: la.flags.scopes,
-		})
-
+		res := contracts.LoginResult{}
 		if err != nil {
-			log.Printf("failed to check auth status: %s", err.Error())
 			res.Status = contracts.LoginStatusUnauthenticated
 		} else {
 			res.Status = contracts.LoginStatusSuccess
 			res.ExpiresOn = &token.ExpiresOn
 		}
-	}
 
-	if !la.flags.onlyCheckStatus && res.Status == contracts.LoginStatusSuccess {
-		// Rehydrate or clear the account's subscriptions cache.
-		// The caching is done here to increase responsiveness of listing subscriptions (during azd init).
-		// It also allows an implicit command for the user to refresh cached subscriptions.
-		if la.flags.clientID == "" {
-			// Deleting subscriptions on file is very unlikely to fail, unless there are serious filesystem issues.
-			// If this does fail, we want the user to be aware of this. Like other stored azd account data,
-			// stored subscriptions are currently tied to the OS user, and not the individual account being logged in,
-			// this means cross-contamination is possible.
-			if err := la.accountSubManager.ClearSubscriptions(ctx); err != nil {
-				return nil, err
-			}
-
-			err := la.accountSubManager.RefreshSubscriptions(ctx)
-			if err != nil {
-				// If this fails, the subscriptions will still be loaded on-demand.
-				// erroring out when the user interacts with subscriptions is much more user-friendly.
-				log.Printf("failed retrieving subscriptions: %v", err)
-			}
+		if la.formatter.Kind() != output.NoneFormat {
+			return nil, la.formatter.Format(res, la.writer, nil)
 		} else {
-			// Service principals do not typically require subscription caching (running in CI scenarios)
-			// We simply clear the cache, which is much faster than rehydrating.
-			err := la.accountSubManager.ClearSubscriptions(ctx)
-			if err != nil {
-				log.Printf("failed clearing subscriptions: %v", err)
+			var msg string
+			switch res.Status {
+			case contracts.LoginStatusSuccess:
+				msg = cLoginSuccessMessage
+			case contracts.LoginStatusUnauthenticated:
+				msg = "Not logged in, run `azd auth login` to login to Azure."
+			default:
+				panic("Unhandled login status")
 			}
+
+			fmt.Fprintln(la.console.Handles().Stdout, msg)
+			return nil, nil
 		}
 	}
 
-	if la.formatter.Kind() == output.NoneFormat {
-		if res.Status == contracts.LoginStatusSuccess {
-			fmt.Fprintln(la.console.Handles().Stdout, "Logged in to Azure.")
-		} else {
-			fmt.Fprintln(la.console.Handles().Stdout, "Not logged in, run `azd auth login` to login to Azure.")
-		}
-
-		return nil, nil
+	if err := la.accountSubManager.ClearSubscriptions(ctx); err != nil {
+		log.Printf("failed clearing subscriptions: %v", err)
 	}
 
-	return nil, la.formatter.Format(res, la.writer, nil)
+	if err := la.login(ctx); err != nil {
+		return nil, err
+	}
+
+	if _, err := la.verifyLoggedIn(ctx); err != nil {
+		return nil, err
+	}
+
+	// Rehydrate or clear the account's subscriptions cache.
+	// The caching is done here to increase responsiveness of listing subscriptions (during azd init).
+	// It also allows an implicit command for the user to refresh cached subscriptions.
+	if la.flags.clientID == "" {
+		// Deleting subscriptions on file is very unlikely to fail, unless there are serious filesystem issues.
+		// If this does fail, we want the user to be aware of this. Like other stored azd account data,
+		// stored subscriptions are currently tied to the OS user, and not the individual account being logged in,
+		// this means cross-contamination is possible.
+		if err := la.accountSubManager.ClearSubscriptions(ctx); err != nil {
+			return nil, err
+		}
+
+		err := la.accountSubManager.RefreshSubscriptions(ctx)
+		if err != nil {
+			// If this fails, the subscriptions will still be loaded on-demand.
+			// erroring out when the user interacts with subscriptions is much more user-friendly.
+			log.Printf("failed retrieving subscriptions: %v", err)
+		}
+	} else {
+		// Service principals do not typically require subscription caching (running in CI scenarios)
+		// We simply clear the cache, which is much faster than rehydrating.
+		err := la.accountSubManager.ClearSubscriptions(ctx)
+		if err != nil {
+			log.Printf("failed clearing subscriptions: %v", err)
+		}
+	}
+
+	la.console.Message(ctx, cLoginSuccessMessage)
+	return nil, nil
+}
+
+// Verifies that the user has credentials stored,
+// and that the credentials stored is accepted by the identity server (can be exchanged for access token).
+func (la *loginAction) verifyLoggedIn(ctx context.Context) (*azcore.AccessToken, error) {
+	credOptions := auth.CredentialForCurrentUserOptions{
+		TenantID: la.flags.tenantID,
+	}
+
+	cred, err := la.authManager.CredentialForCurrentUser(ctx, &credOptions)
+	if err != nil {
+		return nil, err
+	}
+
+	// Ensure credential is valid, and can be exchanged for an access token
+	token, err := cred.GetToken(ctx, policy.TokenRequestOptions{
+		Scopes: la.flags.scopes,
+	})
+
+	if err != nil {
+		return nil, err
+	}
+
+	return &token, nil
 }
 
 func countTrue(elms ...bool) int {