+++ cli/azd/pkg/auth/cache_test.go
@@ -4,14 +4,16 @@
 package auth
 
 import (
+	"context"
 	"testing"
 
+	"github.com/AzureAD/microsoft-authentication-library-for-go/apps/cache"
 	"github.com/stretchr/testify/require"
 )
 
 func TestCache(t *testing.T) {
 	root := t.TempDir()
-
+	ctx := context.Background()
 	c := newCache(root)
 
 	d1 := fixedMarshaller{
@@ -23,15 +25,19 @@ func TestCache(t *testing.T) {
 	}
 
 	// write some data.
-	c.Export(&d1, "d1")
-	c.Export(&d2, "d2")
+	err := c.Export(ctx, &d1, cache.ExportHints{PartitionKey: "d1"})
+	require.NoError(t, err)
+	err = c.Export(ctx, &d2, cache.ExportHints{PartitionKey: "d2"})
+	require.NoError(t, err)
 
 	var r1 fixedMarshaller
 	var r2 fixedMarshaller
 
 	// read back that data we wrote.
-	c.Replace(&r1, "d1")
-	c.Replace(&r2, "d2")
+	err = c.Replace(ctx, &r1, cache.ReplaceHints{PartitionKey: "d1"})
+	require.NoError(t, err)
+	err = c.Replace(ctx, &r2, cache.ReplaceHints{PartitionKey: "d2"})
+	require.NoError(t, err)
 
 	require.NotNil(t, r1.val)
 	require.NotNil(t, r2.val)
@@ -41,13 +47,24 @@ func TestCache(t *testing.T) {
 	// the data should be shared across instances.
 	c = newCache(root)
 
-	c.Replace(&r1, "d1")
-	c.Replace(&r2, "d2")
+	err = c.Replace(ctx, &r1, cache.ReplaceHints{PartitionKey: "d1"})
+	require.NoError(t, err)
+	err = c.Replace(ctx, &r2, cache.ReplaceHints{PartitionKey: "d2"})
+	require.NoError(t, err)
 
 	require.NotNil(t, r1.val)
 	require.NotNil(t, r2.val)
 	require.Equal(t, d1.val, r1.val)
 	require.Equal(t, d2.val, r2.val)
+
+	// read some non-existing data
+	nonExist := fixedMarshaller{
+		val: []byte("some data"),
+	}
+	err = c.Replace(ctx, &nonExist, cache.ReplaceHints{PartitionKey: "nonExist"})
+	require.NoError(t, err)
+	// data should not be overwritten when key is not found.
+	require.Equal(t, []byte("some data"), nonExist.val)
 }
 
 func TestCredentialCache(t *testing.T) {
@@ -88,4 +105,8 @@ func TestCredentialCache(t *testing.T) {
 	require.NotNil(t, r2)
 	require.Equal(t, d1, r1)
 	require.Equal(t, d2, r2)
+
+	// read some non-existing data, ensure errCacheKeyNotFound is returned.
+	_, err = c.Read("nonExist")
+	require.ErrorIs(t, err, errCacheKeyNotFound)
 }