+++ cli/azd/pkg/output/ux/list_as_text_test.go
@@ -12,25 +12,33 @@ import (
 func Test_ListAsText(t *testing.T) {
 	t.Run("single", func(t *testing.T) {
 		expected := "foo"
-		result := ListAsText([]string{expected})
+		result := AndListAsText([]string{expected})
 		require.Equal(t, expected, result)
 	})
 
 	t.Run("double", func(t *testing.T) {
 		expected := "foo and bar"
-		result := ListAsText([]string{"foo", "bar"})
+		result := AndListAsText([]string{"foo", "bar"})
 		require.Equal(t, expected, result)
 	})
 
 	t.Run("triple", func(t *testing.T) {
 		expected := "foo, bar and axe"
-		result := ListAsText([]string{"foo", "bar", "axe"})
+		result := AndListAsText([]string{"foo", "bar", "axe"})
 		require.Equal(t, expected, result)
 	})
 
 	t.Run("long", func(t *testing.T) {
 		expected := "foo, bar, axe, x, y and z"
-		result := ListAsText([]string{"foo", "bar", "axe", "x", "y", "z"})
+		result := AndListAsText([]string{"foo", "bar", "axe", "x", "y", "z"})
+		require.Equal(t, expected, result)
+	})
+
+	//only one test is added for OrListAsText as the function
+	//calls the same listAsText like AndListAstTest
+	t.Run("longOr", func(t *testing.T) {
+		expected := "foo, bar, axe, x, y or z"
+		result := OrListAsText([]string{"foo", "bar", "axe", "x", "y", "z"})
 		require.Equal(t, expected, result)
 	})
 }