+++ cli/azd/pkg/project/service_target_containerapp_test.go
@@ -213,4 +213,5 @@ func setupMocksForContainerApps(mockContext *mocks.MockContext) {
 	)
 	mockazsdk.MockContainerAppSecretsList(mockContext, subscriptionId, resourceGroup, appName, secrets)
 	mockazsdk.MockContainerAppUpdate(mockContext, subscriptionId, resourceGroup, appName, containerApp)
+	mockazsdk.MockContainerRegistryTokenExchange(mockContext, subscriptionId, subscriptionId, "REFRESH_TOKEN")
 }